refactor(filter): extracted condition from filterstep
